tidy(api): rename types -> graphql_types
